server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg[binary]>=3.2,<3.3
sqlalchemy>=2.0.44,<2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1
pydantic>=2.9.2,<3.0.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2
boto3==1.34.14
pillow==10.1.0
spotipy==2.23.0
requests==2.31.0